+++ changelog.md
@@ -8,6 +8,13 @@
 - Add `FormSubmitInteraction.iter_custom_ids_and_values`.
 - Add `FormSubmitInteractionOption.iter_custom_ids_and_values`.
 
+
+##### ext.asyncio
+
+- Implement `Semaphore`. (winwinwinwin#0001)
+- Implement `BoundedSemaphore`. (winwinwinwin#0001)
+
+
 ## 1.1.121 *\[2021-11-28\]*
 
 #### Improvements